Use the jquery-rails-ui gem, just including the modules we need
As far as I can tell, we only use the 'tabs' module in
admin and 'datepicker' on the user-facing part of the site.

The advantage of using this packaging of the gem is that
its assets are in the gem, which simplifies things greatly -
otherwise we'd end up doing something like rewrite the jquery-ui
CSS to SCSS, referencing the image assets via sass-rails helpers
or keep them in their expected paths in public or something.

(Thanks to Louise Crow for pointing out the problem of just
moving jquery-ui's image assets into the asset pipeline.)